Fixed all -Wformat warnings.

+++ conversion.hpp
@@ -231,8 +231,8 @@ h01_to_phredlike(double h0pos, double h0tot, double h1pos, double h1tot,
         double pseudocount, double err_amp_ratio) {
     assert(h0pos <  h0tot || !fprintf(stderr, "%lf <  %lf failed", h0pos, h0tot));
     assert(h1pos <= h1tot || !fprintf(stderr, "%lf <= %lf failed", h1pos, h1tot));
-    assert(h0pos >  0     || !fprintf(stderr, "%lf >  %lf failed", h0pos, 0));
-    assert(h1pos >= 0     || !fprintf(stderr, "%lf >= %lf failed", h1pos, 0));
+    assert(h0pos >  0     || !fprintf(stderr, "%lf >  %lf failed", h0pos, (double)0));
+    assert(h1pos >= 0     || !fprintf(stderr, "%lf >= %lf failed", h1pos, (double)0));
     if (h1pos <= (DBL_EPSILON * pseudocount)) {
         return 0;
     }
